fix crash when light is only transmitted and not reflected inside spheres

+++ Sphere.cpp
@@ -41,9 +41,9 @@ bool Sphere::solveQuadratic(const float &a, const float &b, const float &c, floa
 {
     float discr = b * b - 4 * a * c;
     if (discr < 0) return false;
-    else if (discr == 0) x0 = x1 = - 0.5 * b / a;
+    else if (discr == 0) x0 = x1 = (float) (- 0.5 * b / a);
     else {
-        float q = (b > 0) ? -0.5 * (b + sqrt(discr)) : -0.5 * (b - sqrt(discr));
+        float q = (float) ((b > 0) ? -0.5 * (b + sqrt(discr)) : -0.5 * (b - sqrt(discr)));
         x0 = q / a;
         x1 = c / q;
     }
